+++ product-service/.gitignore
@@ -1,5 +1,5 @@
 # package directories
-node_modules
+../node_modules
 jspm_packages
 
 # Serverless directories
