+++ product-service/webpack.config.js
@@ -1,0 +1,6 @@
+const slsw = require('serverless-webpack');
+
+module.exports = {
+    entry: slsw.lib.entries,
+    target: 'node',
+};
